Initialize if r10k.yaml is present only

+++ docker-entrypoint.d/05-initialize-code.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-touch /etc/puppetlabs/code/environments/r10k-initializing.lock
-/nss_wrapper.sh g10k -config /etc/puppetlabs/r10k/r10k.yaml -info
-rm /etc/puppetlabs/code/environments/r10k-initializing.lock
+
+if test -f /etc/puppetlabs/r10k/r10k.yaml; then
+  touch /etc/puppetlabs/code/environments/r10k-initializing.lock
+  /nss_wrapper.sh g10k -config /etc/puppetlabs/r10k/r10k.yaml -info
+  rm /etc/puppetlabs/code/environments/r10k-initializing.lock
+fi
